Image, css, static folder, and post
app.js was updated with the correct post method while signup.html was given the bootstrap simple form and the right adjustments to allow it to work with app.js. The static method was added so the website could pull from a static public folder, currently only has styles.css and an image.